refactor(backend): Member 도메인 위치 이동 — auth → member 패키지
- auth/domain/Member.java → member/domain/Member.java (git mv)
- auth/repository/MemberRepository.java → member/repository/MemberRepository.java (git mv)
- 5 의존자 import 갱신: AuthService, CurrentMemberArgumentResolver,
  AuthServiceTest, GlobalExceptionHandlerTest
- Member의 근본 책임은 회원 정보 관리 — auth 도메인 종속 제거,
  Step 5 schedule / 이상진 push 등 다른 도메인이 자연스럽게 import 가능
- 비즈니스 메서드 추가는 별 commit (feat)
