+++ backend/src/main/java/com/todayway/backend/auth/service/AuthService.java
@@ -1,13 +1,13 @@
 package com.todayway.backend.auth.service;
 
-import com.todayway.backend.auth.domain.Member;
 import com.todayway.backend.auth.domain.RefreshToken;
+import com.todayway.backend.member.domain.Member;
 import com.todayway.backend.auth.dto.LoginRequest;
 import com.todayway.backend.auth.dto.LoginResponse;
 import com.todayway.backend.auth.dto.SignupRequest;
 import com.todayway.backend.auth.dto.SignupResponse;
-import com.todayway.backend.auth.repository.MemberRepository;
 import com.todayway.backend.auth.repository.RefreshTokenRepository;
+import com.todayway.backend.member.repository.MemberRepository;
 import com.todayway.backend.common.exception.BusinessException;
 import com.todayway.backend.common.exception.ErrorCode;
 import com.todayway.backend.common.jwt.JwtProperties;
